24/02/2025

+++ EjemplosGeneticos/src/us/lsi/expression/DatosExpression.java
@@ -86,7 +86,7 @@ public class DatosExpression implements ExpressionData {
 
 	@Override
 	public Type constType() {
-		return Type.Int;
+		return Type.Double;
 	}
 
 }
